refactor(tui): split the MultiPaneTui impl by concern
Key handling, mouse handling, scrollback, and the agents overlay each
get their own file under tui::multi_pane; the module root keeps the
struct, construction, snapshot application, and geometry.
